requirements.txt
fastapi
uvicorn
yt-dlp
ffmpeg-python
redis
celery
tenacity
pydantic
firebase-admin
python-multipart
python-dotenv
aiofiles
anyio
